added requested  changes
